+++ FineTune/Views/Rows/InactiveAppRow.swift
@@ -165,6 +165,7 @@ struct InactiveAppRow: View {
                             // Unmute: restore to default if at 0
                             if sliderValue == 0 {
                                 sliderValue = defaultUnmuteVolume
+                                onVolumeChange(VolumeMapping.sliderToGain(defaultUnmuteVolume))
                             }
                             onMuteChange(false)
                         } else {
@@ -184,6 +185,7 @@ struct InactiveAppRow: View {
                     .frame(width: DesignTokens.Dimensions.sliderWidth)
                     .opacity(showMutedIcon ? 0.5 : 1.0)
                     .onChange(of: sliderValue) { _, newValue in
+                        guard isEditing else { return }
                         let gain = VolumeMapping.sliderToGain(newValue)
                         onVolumeChange(gain)
                     }